Trait oak_remote_attestation_interactive::handshaker::AttestationVerifier

pub trait AttestationVerifier: Clone + Send + Sync {
    fn verify_attestation(
        &self,
        attestation: &[u8],
        expected_attested_data: &[u8]
    ) -> Result<()>;
}

A trait implementing the functionality of verifying a remote attestation report.

An implementation of this trait is not expected to run in a TEE (i.e. it is usually in the client).

Required Methods

fn verify_attestation(
&self,
attestation: &[u8],
expected_attested_data: &[u8]
) -> Result<()>

Verify the provided remote attestation report, checking that expected_attested_data is cryptographically bound to it (e.g. via a signature).

That is usually generated by AttestationGenerator::generate_attestation.