Module oak_remote_attestation_interactive::handshaker

Structs

AttestationBehavior

Defines the behavior of the remote attestation protocol.

ClientHandshaker

Client of the remote attestation protocol handshake.

EmptyAttestationGenerator

An instance of AttestationGenerator that always returns an empty attestation.

EmptyAttestationVerifier

An instance of AttestationVerifier that succeeds iff the provided attestation is empty.

Encryptor

Used by the client and server to encrypt and decrypt messages with the keys established in the remote attestation protocol handshake.

ServerHandshaker

Server of the remote attestation protocol handshake.

Traits

AttestationGenerator

A trait implementing the functionality of generating a remote attestation report.

AttestationVerifier

A trait implementing the functionality of verifying a remote attestation report.

Functions

attestation_data

Compute data to be attested based on the actual or expected parameters.

hash_concat_hash

Compute a hash over values of possibly different length.